refactor: split transaction file
