+++ base_layer/core/src/transactions/transaction/transaction_builder.rs
@@ -1,0 +1,117 @@
+//  Copyright 2021. The Tari Project
+//
+//  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
+//  following conditions are met:
+//
+//  1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following
+//  disclaimer.
+//
+//  2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
+//  following disclaimer in the documentation and/or other materials provided with the distribution.
+//
+//  3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
+//  products derived from this software without specific prior written permission.
+//
+//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
+//  INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+//  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+//  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+//  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+//  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+//  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+use tari_common_types::types::BlindingFactor;
+
+use crate::transactions::{
+    aggregated_body::AggregateBody,
+    tari_amount::MicroTari,
+    transaction::{Transaction, TransactionError, TransactionInput, TransactionKernel, TransactionOutput},
+    CryptoFactories,
+};
+
+//----------------------------------------  Transaction Builder   ----------------------------------------------------//
+pub struct TransactionBuilder {
+    body: AggregateBody,
+    offset: Option<BlindingFactor>,
+    script_offset: Option<BlindingFactor>,
+    reward: Option<MicroTari>,
+}
+
+impl TransactionBuilder {
+    /// Create an new empty TransactionBuilder
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    /// Update the offset of an existing transaction
+    pub fn add_offset(&mut self, offset: BlindingFactor) -> &mut Self {
+        self.offset = Some(offset);
+        self
+    }
+
+    /// Update the script offset of an existing transaction
+    pub fn add_script_offset(&mut self, script_offset: BlindingFactor) -> &mut Self {
+        self.script_offset = Some(script_offset);
+        self
+    }
+
+    /// Add an input to an existing transaction
+    pub fn add_input(&mut self, input: TransactionInput) -> &mut Self {
+        self.body.add_input(input);
+        self
+    }
+
+    /// Add an output to an existing transaction
+    pub fn add_output(&mut self, output: TransactionOutput) -> &mut Self {
+        self.body.add_output(output);
+        self
+    }
+
+    /// Moves a series of inputs to an existing transaction, leaving `inputs` empty
+    pub fn add_inputs(&mut self, inputs: &mut Vec<TransactionInput>) -> &mut Self {
+        self.body.add_inputs(inputs);
+        self
+    }
+
+    /// Moves a series of outputs to an existing transaction, leaving `outputs` empty
+    pub fn add_outputs(&mut self, outputs: &mut Vec<TransactionOutput>) -> &mut Self {
+        self.body.add_outputs(outputs);
+        self
+    }
+
+    /// Set the kernel of a transaction. Currently only one kernel is allowed per transaction
+    pub fn with_kernel(&mut self, kernel: TransactionKernel) -> &mut Self {
+        self.body.set_kernel(kernel);
+        self
+    }
+
+    pub fn with_reward(&mut self, reward: MicroTari) -> &mut Self {
+        self.reward = Some(reward);
+        self
+    }
+
+    /// Build the transaction.
+    pub fn build(self, factories: &CryptoFactories) -> Result<Transaction, TransactionError> {
+        if let (Some(script_offset), Some(offset)) = (self.script_offset, self.offset) {
+            let (i, o, k) = self.body.dissolve();
+            let tx = Transaction::new(i, o, k, offset, script_offset);
+            tx.validate_internal_consistency(true, factories, self.reward)?;
+            Ok(tx)
+        } else {
+            Err(TransactionError::ValidationError(
+                "Transaction validation failed".into(),
+            ))
+        }
+    }
+}
+
+impl Default for TransactionBuilder {
+    fn default() -> Self {
+        Self {
+            offset: None,
+            body: AggregateBody::empty(),
+            reward: None,
+            script_offset: None,
+        }
+    }
+}
